+++ changelog.md
@@ -9,3 +9,5 @@ Config the System version in **init**.py
 - Initial Create: initial created on December 3, 2022 -> **0.01**
 
 - Initial Development#1: application rules implemented on December 3, 2022 -> **0.1**
+
+- Initial Development#2: database schema implemented and initial developed on December 4, 2022 -> **0.2**